% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staccuracy.R
\name{sa_diff}
\alias{sa_diff}
\title{Statistical tests for the differences between standardized accuracies (staccuracies)}
\usage{
sa_diff(
  actual,
  preds,
  ...,
  na.rm = FALSE,
  sa = NULL,
  pct = c(0.01, 0.02, 0.03, 0.04, 0.05),
  boot_alpha = 0.05,
  boot_it = 1000,
  seed = 0
)
}
\arguments{
\item{actual}{numeric vector. The actual (true) labels.}

\item{preds}{named list of at least two numeric vectors. Each element is a vector of the same length as actual with predictions for each row corresponding to each element of actual. The names of the list elements will be used for the columns of the result.}

\item{...}{not used. Forces explicit naming of subsequent arguments.}

\item{na.rm}{See documentation for \code{\link[=staccuracy]{staccuracy()}}}

\item{sa}{list of functions. Each element is the unquoted name of a staccuracy function. If an element is named, the name will be displayed as the value of the \code{sa} column of the result. Otherwise, the function name will be displayed. If NULL (default), staccuracy functions will be automatically selected based on the datatypes of actual and \code{preds}.}

\item{pct}{numeric with values from (0, 1). The percentage values on which the difference in staccuracies will be tested.}

\item{boot_alpha}{numeric(1) from 0 to 1. Alpha for percentile-based confidence interval range for the bootstrapped means; the bootstrap confidence intervals will be the lowest and highest \code{(1 - 0.05) / 2} percentiles. For example, if \code{boot_alpha = 0.05} (default), the intervals will be at the 2.5 and 97.5 percentiles.}

\item{boot_it}{positive integer(1). The number of bootstrap iterations.}

\item{seed}{integer(1). Random seed for the bootstrap sampling. Supply this between runs to assure identical results.}
}
\value{
tibble. Columns are \code{sa} (name of staccuracy measure), then a column for each named element in \code{preds} (staccuracy for that prediction), then a column for each element of \code{pct} (p-value; percentage of iterations where the difference between staccuracy in the pair of \code{preds} is equal or greater than the \code{pct} value). E.g., for the default \code{pct = c(0.01, 0.02, 0.03, 0.04, 0.05)}, these columns would be \code{p_1}, \code{p_2}, \code{p_3}, \code{p_4}, and  \code{p_5}. When there are more than two predictions, then each row will compare only two at a time.
}
\description{
Because the distribution of staccuracies is uncertain (and indeed, different staccuracies likely have different distributions), bootstrapping is used to empirically estimate the distributions and calculate the p-values.
}
\examples{
lm_attitude_all <- lm(rating ~ ., data = attitude)
lm_attitude__c <- lm(rating ~ . - complaints, data = attitude)

sa_diff(
  attitude$rating,
  list(
    all = predict(lm_attitude_all),
    madv = predict(lm_attitude__c)
  ),
  boot_it = 10
)

}
