% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staccuracy-package.R
\docType{package}
\name{staccuracy-package}
\alias{staccuracy-package}
\title{Standardized Accuracy and Other Model Performance Metrics}
\description{
Standardized accuracy (staccuracy) is a framework for expressing accuracy scores such that 50\% represents a reference level of performance and 100\% is perfect prediction. The 'staccuracy' package provides tools for creating staccuracy functions as well as some recommended staccuracy measures. It also provides functions for some classic performance metrics such as mean absolute error (MAE), root mean squared error (RMSE), and area under the receiver operating characteristic curve (AUCROC), as well as their winsorized versions when applicable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tripartio/staccuracy}
  \item \url{https://tripartio.github.io/staccuracy/}
  \item Report bugs at \url{https://github.com/tripartio/staccuracy/issues}
}

}
\author{
Chitu Okoli \email{Chitu.Okoli@skema.edu}
}
\keyword{internal}
